requirements.txt
markdown==3.4.4
weasyprint==59.0
PyYAML==6.0.1